sales_service/requirements.txt
flask
mysql-connector-python
pytest
pytest-mock
memory_profiler
coverage
jwt
pyotp